changes from review comments
