+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/DraftOrdersController.java
@@ -32,7 +32,7 @@ import uk.gov.hmcts.reform.fpl.service.ValidateGroupService;
 import uk.gov.hmcts.reform.fpl.service.ccd.CoreCaseDataService;
 import uk.gov.hmcts.reform.fpl.service.docmosis.StandardDirectionOrderGenerationService;
 import uk.gov.hmcts.reform.fpl.service.time.Time;
-import uk.gov.hmcts.reform.fpl.utils.CaseDataConverter;
+import uk.gov.hmcts.reform.fpl.utils.CaseConverter;
 import uk.gov.hmcts.reform.fpl.validation.groups.DateOfIssueGroup;
 
 import java.time.LocalDate;
@@ -71,13 +71,13 @@ public class DraftOrdersController {
     private final Time time;
     private final ValidateGroupService validateGroupService;
     private final StandardDirectionsService standardDirectionsService;
-    private final CaseDataConverter caseDataConverter;
+    private final CaseConverter caseConverter;
 
     private static final String JUDGE_AND_LEGAL_ADVISOR_KEY = "judgeAndLegalAdvisor";
 
     @PostMapping("/about-to-start")
     public AboutToStartOrSubmitCallbackResponse handleAboutToStart(@RequestBody CallbackRequest callbackRequest) {
-        CaseData caseData = caseDataConverter.convertToCaseData(callbackRequest);
+        CaseData caseData = caseConverter.convertToCaseData(callbackRequest.getCaseDetails());
 
         LocalDate dateOfIssue = time.now().toLocalDate();
         Order standardDirectionOrder = caseData.getStandardDirectionOrder();
@@ -91,14 +91,14 @@ public class DraftOrdersController {
             .build();
 
         return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(caseDataConverter.convertToMap(updatedCaseData))
+            .data(caseConverter.convertToMap(updatedCaseData))
             .build();
     }
 
     @PostMapping("/date-of-issue/mid-event")
     public AboutToStartOrSubmitCallbackResponse handleMidEventIssueDate(@RequestBody CallbackRequest callbackRequest) {
         CaseDetails caseDetails = callbackRequest.getCaseDetails();
-        CaseData caseData = caseDataConverter.convertToCaseData(caseDetails);
+        CaseData caseData = caseConverter.convertToCaseData(caseDetails);
 
         List<String> errors = validateGroupService.validateGroup(caseData, DateOfIssueGroup.class);
 
@@ -131,7 +131,7 @@ public class DraftOrdersController {
 
     @PostMapping("/mid-event")
     public AboutToStartOrSubmitCallbackResponse handleMidEvent(@RequestBody CallbackRequest callbackRequest) {
-        CaseData caseData = caseDataConverter.convertToCaseData(callbackRequest);
+        CaseData caseData = caseConverter.convertToCaseData(callbackRequest.getCaseDetails());
 
         JudgeAndLegalAdvisor judgeAndLegalAdvisor = getSelectedJudge(caseData.getJudgeAndLegalAdvisor(),
             caseData.getAllocatedJudge());
@@ -153,14 +153,14 @@ public class DraftOrdersController {
         order.setOrderDocReferenceFromDocument(document);
 
         return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(caseDataConverter.convertToMap(updated))
+            .data(caseConverter.convertToMap(updated))
             .build();
     }
 
     @PostMapping("/about-to-submit")
     public AboutToStartOrSubmitCallbackResponse handleAboutToSubmit(@RequestBody CallbackRequest callbackRequest) {
         CaseDetails caseDetails = callbackRequest.getCaseDetails();
-        CaseData caseData = caseDataConverter.convertToCaseData(caseDetails);
+        CaseData caseData = caseConverter.convertToCaseData(caseDetails);
 
         List<String> validationErrors = orderValidationService.validate(caseData);
         if (!validationErrors.isEmpty()) {
@@ -206,13 +206,13 @@ public class DraftOrdersController {
         order.setOrderDocReferenceFromDocument(document);
 
         return AboutToStartOrSubmitCallbackResponse.builder()
-            .data(caseDataConverter.convertToMap(updated))
+            .data(caseConverter.convertToMap(updated))
             .build();
     }
 
     @PostMapping("/submitted")
     public void handleSubmittedEvent(@RequestBody CallbackRequest callbackRequest) {
-        CaseData caseData = caseDataConverter.convertToCaseData(callbackRequest);
+        CaseData caseData = caseConverter.convertToCaseData(callbackRequest.getCaseDetails());
 
         Order standardDirectionOrder = caseData.getStandardDirectionOrder();
         if (standardDirectionOrder.getOrderStatus() != OrderStatus.SEALED) {
